Update migration table data type
This commit updates the data type for 'user_id' and 'total_records' in the 'create_bulk_actions_table'. The 'user_id' is now a foreignId and 'total_records' is now a bigInteger. In addition, the data type for 'record_id' in the 'create_bulk_action_records_table' was changed from uuid to bigInteger.